// frontend/src/api/generated/models/bookmark-save-json.ts
/* tslint:disable */
/* eslint-disable */
/**
 * chainlink-api API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BookmarkSaveJson
 */
export interface BookmarkSaveJson {
    /**
     * 
     * @type {string}
     * @memberof BookmarkSaveJson
     */
    collectionId: string;
    /**
     * 
     * @type {string}
     * @memberof BookmarkSaveJson
     */
    folderId?: string;
    /**
     * 
     * @type {string}
     * @memberof BookmarkSaveJson
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof BookmarkSaveJson
     */
    url: string;
    /**
     * 
     * @type {string}
     * @memberof BookmarkSaveJson
     */
    description?: string;
    /**
     * 
     * @type {Set<string>}
     * @memberof BookmarkSaveJson
     */
    tagIds?: Set<string>;
}

/**
 * Check if a given object implements the BookmarkSaveJson interface.
 */
export function instanceOfBookmarkSaveJson(value: object): value is BookmarkSaveJson {
    if (!('collectionId' in value) || value['collectionId'] === undefined) return false;
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('url' in value) || value['url'] === undefined) return false;
    return true;
}

export function BookmarkSaveJsonFromJSON(json: any): BookmarkSaveJson {
    return BookmarkSaveJsonFromJSONTyped(json, false);
}

export function BookmarkSaveJsonFromJSONTyped(json: any, ignoreDiscriminator: boolean): BookmarkSaveJson {
    if (json == null) {
        return json;
    }
    return {
        
        'collectionId': json['collectionId'],
        'folderId': json['folderId'] == null ? undefined : json['folderId'],
        'title': json['title'],
        'url': json['url'],
        'description': json['description'] == null ? undefined : json['description'],
        'tagIds': json['tagIds'] == null ? undefined : new Set(json['tagIds']),
    };
}

export function BookmarkSaveJsonToJSON(json: any): BookmarkSaveJson {
    return BookmarkSaveJsonToJSONTyped(json, false);
}

export function BookmarkSaveJsonToJSONTyped(value?: BookmarkSaveJson | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'collectionId': value['collectionId'],
        'folderId': value['folderId'],
        'title': value['title'],
        'url': value['url'],
        'description': value['description'],
        'tagIds': value['tagIds'] == null ? undefined : Array.from(value['tagIds'] as Set<any>),
    };
}